+++ src/Icons/FontAwesome6ProIconTheme.php
@@ -83,6 +83,7 @@ class FontAwesome6ProIconTheme implements IconTheme
         AppShellIcons::SPINNER => 'spinner',
         AppShellIcons::PLUG => 'plug',
         AppShellIcons::UPLOAD => 'cloud-upload',
+        AppShellIcons::DOWNLOAD => 'cloud-download',
         AppShellIcons::FOLDER => 'folder-open',
         AppShellIcons::FILE => 'file',
     ];