+++ src/Icons/FontAwesomeIconTheme.php
@@ -77,6 +77,7 @@ class FontAwesomeIconTheme implements IconTheme
         AppShellIcons::SPINNER => 'spinner',
         AppShellIcons::PLUG => 'plug',
         AppShellIcons::UPLOAD => 'cloud-upload-alt',
+        AppShellIcons::DOWNLOAD => 'cloud-download-alt',
         AppShellIcons::FOLDER => 'folder',
         AppShellIcons::FILE => 'file',
     ];